requirements.txt
pytest==5.4.1
pytest_yaml==1.2.1
allure-pytest==2.8.13
selenium==3.141.0
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
matplotlib==3.3.2
#opencv-python==4.4.0.42
#opencv-contrib-python==4.4.0.44
imutils==0.5.3